evm-rs: EVM-RS support with a dedicated evm-rs build script, architecture and cross-build detection
