+++ build_with_svm.sh
